Move packages into a types-internal pkg
This allows for there to be no API changes from the circular dependencies fix.

There is now a `@parcel/types` package with the same API as before and a new
`@parcel/types-internal` package, which is consumed by `@parcel/workers` while
avoiding it.
