+++ packages/core/integration-tests/test/sourcemaps.js
@@ -17,7 +17,7 @@ import {loadSourceMapUrl} from '@parcel/utils';
 import nullthrows from 'nullthrows';
 import type WorkerFarm from '@parcel/workers';
 
-const bundle = (name, opts?: InitialParcelOptions<WorkerFarm>) => {
+const bundle = (name, opts?: InitialParcelOptions) => {
   return _bundle(
     name,
     mergeParcelOptions(
