+++ packages/core/types-internal/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@parcel/types",
+  "name": "@parcel/types-internal",
   "version": "2.12.0",
   "license": "MIT",
   "main": "src/index.js",
@@ -8,15 +8,15 @@
     "type": "git",
     "url": "https://github.com/parcel-bundler/parcel.git"
   },
-  "publishConfig": {
-    "access": "public"
-  },
+  "publishConfig": {},
   "scripts": {
     "build-ts": "./scripts/build-ts.sh",
     "check-ts": "tsc --noEmit lib/index.d.ts"
   },
   "dependencies": {
-    "@parcel/types-internal": "2.12.0",
-    "@parcel/workers": "2.12.0"
+    "@parcel/diagnostic": "2.12.0",
+    "@parcel/feature-flags": "2.12.0",
+    "@parcel/source-map": "^2.1.1",
+    "utility-types": "^3.10.0"
   }
 }
