+++ packages/core/types-internal/scripts/build-ts.js
