+++ packages/core/types-internal/src/Cache.js
