+++ packages/core/types-internal/src/DependencySpecifier.js
