+++ packages/core/types-internal/src/FileCreateInvalidation.js
