+++ packages/core/types-internal/src/Glob.js
