+++ packages/core/types-internal/src/PackageManager.js
