+++ packages/core/types-internal/src/PackageName.js
